Allow for scoreboard objective to coexist with others; should fix #5
When giving the player a scoreboard, their existing one is used if it
exists. This will allow for other objectives to exist on the scoreboard
in different slots.

If an objective already exists in our desired slot (SIDEBAR), a passive
mode is available to prefer this instead of our own objective.
Otherwise, this objective is replaced.