Fix countdown sign algorithm and resolve bug #1
Countdown signs were not showing the correct time as the value for the
time elapsed was retrieved from the configuration (an old value).

EventEndTask now controls working out the time that has elapsed on an
hourly basis (as this task controls timing for events).

The algorithm for countdown signs is now corrected to properly show the
number of hours, minutes, and seconds.

The correct event (SignChangeEvent) is now handled instead of
BlockPlaceEvent, as this is triggered when the user actually changes a
sign; block place event would only show blank lines on the sign as the
sign starts off blank. This was the cause of issue #1.