Implement point strategies
So-called 'point strategies' are implemented, which allow for the
tracking of arbitrary values instead of just kills.

The main implementations of point strategies are Constant and
TokenRewards. Constant uses a constant value for the number of points so
as to allow for the previous functionality of tracking kills, for
example. TokenRewards uses the module TokenRewards to determine a value
of points and uses this value instead.